requirements.txt
matplotlib>=3.0
paho-mqtt>=1.6
requests>=2.25.0
